requirements.txt
anyio==4.7.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.2
fastapi==0.95.0
filelock==3.16.1
fsspec==2024.12.0
h11==0.14.0
huggingface-hub==0.27.0
idna==3.10
imageio==2.36.1
Jinja2==3.1.4
lazy_loader==0.4
MarkupSafe==3.0.2
mpmath==1.3.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.0
opencv-python-headless==4.10.0.84
packaging==24.2
pillow==11.0.0
pydantic==1.10.19
pyngrok==7.2.2
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
scikit-image==0.25.0
scipy==1.14.1
sniffio==1.3.1
starlette==0.26.1
sympy==1.13.1
tifffile==2024.12.12
tokenizers==0.21.0
torch==2.5.1
tqdm==4.67.1
transformers==4.47.1
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.1